fix: position type 파일명 refactor 이후 잘못된 경로 수정
